Implement pam_sm_authenticate, preare for the rest
Simple pm_sm_authenticate that implements asking for a
password if we have none, and prepares the code to either
wait for pam_sm_open_session or directly unlock the kwallet.

Unlocking the kwallet in the pam_sm_authenticate is needed in
the case where the screenlocker closes the wallet.