fix off by one in the socket filename
from the snprintf man page:

The functions snprintf() and vsnprintf() do not write more than size bytes (*including* the terminating null byte  ('\0')).   If
the  output was truncated due to this limit, then the return value is the number of characters (*excluding* the terminating null
byte) which would have been written to the final string if enough space had been available.  Thus, a return value of  size  or
more means that the output was truncated.

pam_wallet.c should increase the needed parameter by one, to account for the terminating null byte  ('\0')

REVIEW: 129955