Make kwallet-pam work with pam_fscrypt
Summary:
fscrypt encrypts folders, you can use pam_fscrypt to auto-unencrypt
some, for example your home folder.

This means that the old code of kwallet-pam would not work
since it tries to read/write the salt file at the auth step.

This change moves that from the auth step to the open step, sadly that
means we need to keep the password in (pam "secure") memory for a while more.

Test Plan: I can now login into my session and the kwallet password prompt is not shown

Reviewers: sitter

Reviewed By: sitter

Subscribers: sitter, security-team, davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27935